Updates to improve store
