+++ src/state/hooks.ts
@@ -1,0 +1,4 @@
+import { useContext } from "react";
+import { SettingContext } from "./Context";
+
+export const useSetting = () => useContext(SettingContext);
