+++ src/state/model.ts
@@ -1,0 +1,9 @@
+import { backgroundImgUrl } from "../settingNames";
+
+export type State = {
+  backgroundImage: string;
+};
+
+export const initialState: State = {
+  backgroundImage: localStorage.getItem(backgroundImgUrl) ?? "",
+};
